use webpack.updateConfig.finally to add webpack plugins
