+++ package.json
@@ -33,7 +33,7 @@
     "babel-preset-stage-0": "^6.5.0",
     "coveralls": "^2.11.6",
     "dora": "0.3.x",
-    "dora-plugin-webpack": "0.7.x",
+    "dora-plugin-webpack": "0.6.x",
     "eslint": "^2.9.0",
     "eslint-config-airbnb": "^9.0.1",
     "eslint-plugin-import": "^1.10.2",
