+++ test/fixtures/normal/mod.js
@@ -1,1 +1,1 @@
-console.log('935448');
+console.log('550433');
